+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 1.1.0 / 2019-05-22
+* [ENHANCEMENT] Format go code
+* [ENHANCEMENT] Provide grafana scripts for exporting datasources and dashboards
+
 ## 1.0.0 / 2019-05-20
 * [INITIAL] Initial commit of grafana-config-controller